decompose drag/drop intializer
it's becoming necessary to get a summary block back for manipulation and the old system just made it and placed it. This made it difficult to get it back for further work.

This code refactors the code that puts d/d on blocks. that was the only part of block prep that didn't work on the loose node.

Now the old use of the block builder need to be rewritten to take care of placement of the blocks